requirements.txt
openai
streamlit
ipython
PyYAML
python-box
